requirements.txt
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp>=3.6.0,<3.8.0
aiohttp-cors==0.7.0
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
botocore==1.35.36
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorful==0.5.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.59.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.10.0
pyparsing==3.2.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-ibm-catalog==0.0.0
qiskit-ibm-runtime==0.33.2
qiskit-serverless==0.14.2
ray==2.39.0
referencing==0.35.1
requests==2.32.3
requests_ntlm==1.3.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
rustworkx==0.15.1
s3fs==2024.10.0
scipy==1.14.1
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
stack-data==0.6.3
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.3
textual==0.86.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yarg==0.1.9
yarl==1.18.0
zipp==3.21.0
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-cors==0.7.0
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
botocore==1.35.36
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorful==0.5.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.59.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.10.0
pyparsing==3.2.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qiskit==1.2.4
qiskit-aer==0.15.1
qiskit-ibm-catalog==0.0.0
qiskit-ibm-runtime==0.33.2
qiskit-serverless==0.14.2
ray==2.39.0
referencing==0.35.1
requests==2.32.3
requests_ntlm==1.3.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
rustworkx==0.15.1
s3fs==2024.10.0
scipy==1.14.1
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
stack-data==0.6.3
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.3
textual==0.86.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yarg==0.1.9
yarl==1.18.0
zipp==3.21.0
